requirements.txt
Flask>=2.3
torch>=2.2
torchvision>=0.17
Pillow>=10.0
timm==1.0.9

